+++ packages/e2e-tests/tests/common/atoms-tests/deploy-collection.ts
@@ -1,13 +1,13 @@
 import type { IRaribleSdk } from "@rarible/sdk/src/domain"
 import type { BlockchainWallet } from "@rarible/sdk-wallet"
-import type { DeployResponse, DeployTokenRequest } from "@rarible/sdk/src/types/nft/deploy/domain"
+import type { CreateCollectionResponse, CreateCollectionRequest } from "@rarible/sdk/src/types/nft/deploy/domain"
 
 /**
  * Deploy new collection, await transaction, check address
  */
 export async function deployCollection(sdk: IRaribleSdk,
 									   wallet: BlockchainWallet,
-									   deployRequest: DeployTokenRequest): Promise<DeployResponse> {
+									   deployRequest: CreateCollectionRequest): Promise<CreateCollectionResponse> {
 	console.log("Deploying collection, deploy_token_request=", deployRequest)
 	const deployResult = await sdk.nft.deploy(deployRequest)
 	await deployResult.tx.wait()