+++ packages/e2e-tests/tests/common/wallet.ts
@@ -1,25 +1,37 @@
-import type { BlockchainWallet } from "@rarible/sdk-wallet"
+import type { BlockchainWallet, TezosWallet } from "@rarible/sdk-wallet"
+import { EthereumWallet, FlowWallet } from "@rarible/sdk-wallet"
 import { Blockchain } from "@rarible/api-client"
 import { initProvider } from "@rarible/sdk/src/sdk-blockchains/ethereum/test/init-providers"
-import { EthereumWallet, FlowWallet } from "@rarible/sdk-wallet"
-import type { TezosWallet } from "@rarible/sdk-wallet"
 import { createTestWallet as createTezosWallet } from "@rarible/sdk/src/sdk-blockchains/tezos/test/test-wallet"
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import fcl from "@onflow/fcl"
 import type { UnionAddress } from "@rarible/types"
 import { toUnionAddress } from "@rarible/types"
 import { testsConfig } from "./config"
+import { EVMBlockchain } from "@rarible/sdk/src/sdk-blockchains/ethereum/common";
 
-export function getEthereumWallet(pk?: string): EthereumWallet {
+export function getEthereumWallet(pk?: string): EthereumWallet<EVMBlockchain> {
 	const { web3, wallet } = initProvider(pk)
 	const ethereum = new Web3Ethereum({
 		web3: web3,
 		from: wallet.getAddressString(),
 	})
-	return new EthereumWallet(ethereum)
+	return new EthereumWallet(ethereum, Blockchain.ETHEREUM)
 }
 
-export function getEthereumWalletBuyer(): EthereumWallet {
+export function getPolygonWallet(pk?: string): EthereumWallet<EVMBlockchain> {
+	const { web3, wallet } = initProvider(pk, {
+		networkId: 80001,
+		rpcUl: "https://rpc-mumbai.maticvigil.com",
+	})
+	const ethereum = new Web3Ethereum({
+		web3: web3,
+		from: wallet.getAddressString(),
+	})
+	return new EthereumWallet(ethereum, Blockchain.POLYGON)
+}
+
+export function getEthereumWalletBuyer(): EthereumWallet<EVMBlockchain> {
 	return getEthereumWallet(testsConfig.variables.ETHEREUM_WALLET_BUYER)
 }
 
@@ -41,6 +53,8 @@ export function getFlowWallet(): FlowWallet {
 export async function getWalletAddress(wallet: BlockchainWallet, withPrefix: boolean = true): Promise<string> {
 	switch (wallet.blockchain) {
 		case Blockchain.ETHEREUM:
+			return (withPrefix ? "ETHEREUM:" : "") + (await wallet.ethereum.getFrom())
+		case Blockchain.POLYGON:
 			return (withPrefix ? "ETHEREUM:" : "") + (await wallet.ethereum.getFrom())
 		case Blockchain.TEZOS:
 			return (withPrefix ? "TEZOS:" : "") + (await wallet.provider.address())