+++ packages/sdk/src/sdk-blockchains/ethereum/common/index.ts
@@ -20,6 +20,7 @@ import type {
 	Part,
 } from "@rarible/ethereum-api-client"
 import type { Order } from "@rarible/ethereum-api-client/build/models"
+import type { EthereumTransaction } from "@rarible/ethereum-provider"
 import type { CurrencyType, RequestCurrency } from "../../../common/domain"
 
 export type EVMBlockchain = Blockchain.ETHEREUM | Blockchain.POLYGON
@@ -27,6 +28,8 @@ export const EVMBlockchains: EVMBlockchain[] = [
 	Blockchain.ETHEREUM,
 	Blockchain.POLYGON,
 ]
+
+export type CreateEthereumCollectionResponse = { tx: EthereumTransaction, address: Address }
 
 export function getEthTakeAssetType(currency: RequestCurrency) {
 	switch (currency["@type"]) {
@@ -148,6 +151,10 @@ export function getEVMBlockchain(network: EthereumNetwork): EVMBlockchain {
 			return Blockchain.ETHEREUM
 		case "ropsten":
 			return Blockchain.ETHEREUM
+		case "dev-ethereum":
+			return Blockchain.ETHEREUM
+		case "dev-polygon":
+			return Blockchain.POLYGON
 		case "rinkeby":
 			return Blockchain.ETHEREUM
 		case "mainnet":