+++ packages/sdk/src/sdk-blockchains/ethereum/sale.test.ts
@@ -1,9 +1,8 @@
-import { awaitAll } from "@rarible/ethereum-sdk-test-common"
+import { awaitAll, deployTestErc20, deployTestErc721 } from "@rarible/ethereum-sdk-test-common"
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import { EthereumWallet } from "@rarible/sdk-wallet"
 import { toContractAddress, toItemId } from "@rarible/types"
-import { deployTestErc20 } from "@rarible/protocol-ethereum-sdk/build/order/contracts/test/test-erc20"
-import { deployTestErc721 } from "@rarible/protocol-ethereum-sdk/build/order/contracts/test/test-erc721"
+import { Blockchain } from "@rarible/api-client"
 import { createRaribleSdk } from "../../index"
 import { LogsLevel } from "../../domain"
 import { initProviders } from "./test/init-providers"
@@ -14,8 +13,8 @@ describe("sale", () => {
 	const { web31, web32, wallet1, wallet2 } = initProviders()
 	const ethereum1 = new Web3Ethereum({ web3: web31 })
 	const ethereum2 = new Web3Ethereum({ web3: web32 })
-	const sdk1 = createRaribleSdk(new EthereumWallet(ethereum1), "e2e", { logs: LogsLevel.DISABLED })
-	const sdk2 = createRaribleSdk(new EthereumWallet(ethereum2), "e2e", { logs: LogsLevel.DISABLED })
+	const sdk1 = createRaribleSdk(new EthereumWallet(ethereum1, Blockchain.ETHEREUM), "e2e", { logs: LogsLevel.DISABLED })
+	const sdk2 = createRaribleSdk(new EthereumWallet(ethereum2, Blockchain.ETHEREUM), "e2e", { logs: LogsLevel.DISABLED })
 
 	const conf = awaitAll({
 		testErc20: deployTestErc20(web31, "Test1", "TST1"),