+++ packages/sdk/src/sdk-blockchains/tezos/index.ts
@@ -13,7 +13,7 @@ import { TezosBurn } from "./burn"
 import { TezosTokenId } from "./token-id"
 import { TezosCancel } from "./cancel"
 import { TezosBalance } from "./balance"
-import { TezosDeploy } from "./deploy"
+import { TezosCreateCollection } from "./create-collection"
 import { TezosCanTransfer } from "./restriction"
 
 export function createTezosSdk(
@@ -25,9 +25,10 @@ export function createTezosSdk(
 	const maybeProvider = getMaybeTezosProvider(wallet?.provider, network)
 	const sellService = new TezosSell(maybeProvider, apis)
 	const mintService = new TezosMint(maybeProvider, apis, network)
-	const balanceService = new TezosBalance(maybeProvider, apis)
+	const balanceService = new TezosBalance(maybeProvider)
 	const bidService = new TezosBid(maybeProvider, apis, balanceService, network)
 	const fillService = new TezosFill(maybeProvider, apis, network)
+	const createCollectionService = new TezosCreateCollection(maybeProvider, network)
 
 	return {
 		nft: {
@@ -35,7 +36,8 @@ export function createTezosSdk(
 			burn: new TezosBurn(maybeProvider, apis, network).burn,
 			transfer: new TezosTransfer(maybeProvider, apis, network).transfer,
 			generateTokenId: new TezosTokenId(maybeProvider, apis).generateTokenId,
-			deploy: new TezosDeploy(maybeProvider, network).deployToken,
+			deploy: createCollectionService.createCollection,
+			createCollection: createCollectionService.createCollection,
 			preprocessMeta: Middlewarer.skipMiddleware(mintService.preprocessMeta),
 		},
 		order: {